Type:fix bug 25608
Description:no check without scale list

RootCause:Null

Solution:adjused process

+++ application/widgets/ScaleWidget.h
@@ -57,7 +57,7 @@ private:
     int         m_nCurrentIndex = -1;           //  当前选项
     int         m_nMaxScale = 1000;
 
-    QList<int> dataList;
+    QList<double> dataList;
     QPointer<DocSheet> m_sheet;
 };
